#46: Reorganized sdk table classes in settings to be cleaner
